% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices.R
\name{isi_calculator}
\alias{isi_calculator}
\title{Insulin Sensitivity Indices Calculator}
\usage{
isi_calculator(data, categories = c("fasting", "ogtt", "adipo"))
}
\arguments{
\item{data}{A dataframe containing the required variables for calculation. This includes insulin
and glucose measurements at fasting (0 min), 30 min, and 120 min post-glucose load, as well as
triglycerides, HDL cholesterol, and other metabolic parameters as needed.}

\item{categories}{A character vector specifying which categories of indices to calculate.
If you have only fasting insulin and glucose values choose "fasting" category
If you also have OGTT based vlaues choose "ogtt" category, if you have 0, 30, 120 it calcualte indices based on these values, if 30 min values are missing then it will only calcualte the indices based on 0 and 120 min values.
If you have lipid values like TG, FFA and HDL-C then choose "adipo" category.
OGTT and adipo categories also use anthropometric value ssuch as age, sex, weight, bmi and waist.
If you want to calculate indices for all categories then use c("fasting", "ogtt", "adipo") as input.

Supported options are "fasting", "ogtt", and "adipo". Specific indices calculated for each category are:
\itemize{
\item{\strong{fasting}:}{
\itemize{
\item{Fasting Insulin Sensitivity: Inversed to represent IS}
\item{Raynaud Index: an IS index}
\item{HOMA-IR_inv: Inversed to represent IS}
\item{FRI: Fasting Insulin Resistance Index: Inversed to represent IS}
\item{QUICKI: Quantitative Insulin Sensitivity Check Index: IS index}
\item{Belfiore basal index : IS index}
\item{Insulin to Glucose Ratio: Inversed to represent IS}
\item{Insulin Sensitivity Index basal :IS index}
\item{Bennett Index : IS index}
\item{HOMA-IR-inv (Revised) : Revised HOMA-IR inversed to represent IS Index}
}
}
\item{\strong{ogtt}:}{
\itemize{
\item{Insulin Sensitivity Index at 120 min}
\item{Gutt Index: Insulin sensitivity based on the OGTT}
\item{Avignon Index at 0 and 120 min: A composite measure considering glucose and insulin levels}
\item{Matsuda Index based on AUC for glucose and insulin at 0, 30, 120 minutes }
\item{Isi_120: Insulin Sensitivity Index 120}
\item{Ig_ratio_120_inv: Insulin to Glucose Ratio at 120 : Inversed to represent IS}
\item{Gutt_index: Gutt Index : IS index}
\item{Avignon_Si0: Avignon Index at 0 min}
\item{Avignon_Si120: Avignon Index at 120 min}
\item{Avignon_Sim: Avignon Index mean}
\item{Modified_stumvoll: Modified Stumvoll Index}
\item{Stumvoll_Demographics: Stumvoll Index with Demographics}
\item{Glu_Auc_Mean: Mean Glucose AUC}
\item{Insu_Auc_Mean: Mean Insulin AUC}
\item{Matsuda_Auc: Matsuda Index}
\item{BigttSi: BIGTT-Si}
\item{Ifc_inv: Insulin fold change: Inversed to represent IS}
\item{HIRI_inv: Hepatic Insulin Resistance Index: Inversed to represent IS}
}
}
\item{\strong{adipo}:}{
\itemize{
\item{Revised_QUICKI: Revised QUICK Index}
\item{VAI_Men_inv: Visceral Adiposity Index for Men: Inversed to represent IS}
\item{VAI_Women_inv: Visceral Adiposity Index for Women : Inversed to represent IS}
\item{TG_HDL_C_inv: TG to HDL-C ratio converted to IS :Inversed to represent IS}
\item{TyG_inv: TyG Index converted to IS : Inversed to represent IS}
\item{LAP_Men_inv: Lipid Accumulation Product for Men : Inversed to represent IS}
\item{LAP_Women_inv: Lipid Accumulation Product for Women : Inversed to represent IS}
\item{McAuley_index: McAuley Index}
\item{LIRI_inv: Liver Insulin Resistance Index converted to IS : Inversed to represent IS}
\item{Adipo_inv: Adipose Insulin Resistance Index converted to IS : Inversed to represent IS}
\item{Lipo_inv: Lipolysis Index converted to IS : Inversed to represent IS}
\item{ATIRI_inv: Adipose Tissue Insulin Resistance Index converted to IS : Inversed to represent IS}
\item{Belfiore_inv_FFA: Belfiore Index for FFA converted to IS : Inversed to represent IS}
}
}
}

The calculation of most indices follows established formulas documented in the references, with units and other details conforming to the standards set forth in the literature. Although not all original references are explicitly provided, they were consulted individually for each index calculation.

References:
\itemize{
\item Amalia Gastaldelli (2022). (PubMed: \url{https://onlinelibrary.wiley.com/doi/10.1002/oby.23503})
\item Carlos Lorenzo (2010). (PubMed: \url{https://academic.oup.com/jcem/article/95/11/5082/2835314})
}}
}
\value{
A data frame containing the IS indices calculated for the specified category/ies.
The values are returned raw without any kind of normalization or transformation.
For further processing especially the statistical testing and visualization, it is recommended to normalize the values.
As most of them are on different scales
}
\description{
Calculates surrogate insulin sensitivity indices based on fasting, OGTT, and lipid (adipo) values values.
}
\examples{
data(example_data)
# Example usage of the isi_calculator function
# Run the isi_calculator function with the sample data
# run for each category separately
result <- isi_calculator(example_data, categories = "fasting")
result <- isi_calculator(example_data, categories = "ogtt")
result <- isi_calculator(example_data, categories = "adipo")
# OR all three together
result <- isi_calculator(example_data, categories = c("adipo", "ogtt", "fasting"))
# View the results
print(result)
# use ?isi_calculator to see the full documentation of the function
# use ?example_data to see the sample data description

}
\keyword{internal}
