% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indices.R
\name{isi_calculator}
\alias{isi_calculator}
\title{Insulin Sensitivity Indices Calculator}
\usage{
isi_calculator(data, category = c("fasting", "ogtt", "adipo", "tracer_dxa"))
}
\arguments{
\item{data}{A dataframe containing the required variables for calculation. This includes insulin
and glucose measurements at fasting (0 min), 30 min, and 120 min post-glucose load, as well as
triglycerides, HDL cholesterol, and other metabolic parameters as needed.
Name of the variables in the input data should be the same as the ones listed below in details (or check with ?example_data) for accurately calculating the indices.
Otherwise it will result in Error.
If a variable is missing for the category it will not calculate any of the index for that category.
This can be handled by creating the missing variable column but with simulated, random or "NA" values
If variable column exist but the values are missing \code{isi_calculator} function will set the value to NA in internal operation and calculate the remaining indices while returning the NA value for indices where the the missing variable was required.}

\item{category}{A character vector specifying which category of indices
to calculate. If you have only fasting insulin and glucose values choose
"fasting" category. If you also have OGTT based values choose "ogtt" category;
it calculates indices based on these values, if 30 min values are missing
then it will only calculate the indices based on 0 and 120 min values.
If you have lipid values like TG, FFA, and HDL-C then choose "adipo" category.
OGTT and adipo category also use anthropometric values such as age, sex, weight, bmi, and waist.
If you want to calculate indices for all categories then leave the argument empty or use the a list of category like c("fasting", "ogtt", "adipo", "tracer_dxa").
c("fasting", "ogtt", "adipo") as input.}
}
\value{
A data frame containing the IS indices calculated for the specified category/ies.
The values are returned raw without any kind of normalization or transformation.
For further processing especially the statistical testing and visualization, it is recommended to normalize the values.
As most of them are on different scales
}
\description{
Calculates surrogate insulin sensitivity indices based on fasting, OGTT, and lipid (adipo) values values.
}
\details{
The function requires specific columns in the data for each category:
\itemize{
\item \code{fasting}: \code{"G0", "I0"}
\item \code{ogtt}: \code{"G0", "I0", "G120", "I120", "G30", "I30", "age", "sex", "bmi", "weight"}
\item \code{adipo}: \code{"G0", "I0", "G120", "I120", "G30", "I30", "age", "sex", "bmi", "weight", "TG", "HDL_c", "FFA", "waist"}
\item \code{tracer_dxa}: This category includes all of the columns required for \code{adipo}
plus specific tracer and DXA measures: \code{"rate_palmitate", "rate_glycerol", "fat_mass"}.
Ensure that the data frame contains these columns when selecting this category for accurate calculation.
}

Supported options for \code{category} are "fasting", "ogtt", "adipo", and "tracer_dxa".
Specific indices calculated for each category are detailed within each category section.

\itemize{
\item{\strong{fasting}:}{
\itemize{
\item{Fasting Insulin Sensitivity: Inversed to represent IS}
\item{Raynaud Index: an IS index}
\item{HOMA-IR_inv: Inversed to represent IS}
\item{FRI: Fasting Insulin Resistance Index: Inversed to represent IS}
\item{QUICKI: Quantitative Insulin Sensitivity Check Index: IS index}
\item{Belfiore basal index : IS index}
\item{Insulin to Glucose Ratio: Inversed to represent IS}
\item{Insulin Sensitivity Index basal :IS index}
\item{Bennett Index : IS index}
\item{HOMA-IR-inv (Revised) : Revised HOMA-IR inversed to represent IS Index}
}
}
\item{\strong{ogtt}:}{
\itemize{
\item{Insulin Sensitivity Index at 120 min}
\item{Gutt Index: Insulin sensitivity based on the OGTT}
\item{Avignon Index at 0 and 120 min: A composite measure considering glucose and insulin levels}
\item{Matsuda Index based on AUC for glucose and insulin at 0, 30, 120 minutes }
\item{Isi_120: Insulin Sensitivity Index 120}
\item{Ig_ratio_120_inv: Insulin to Glucose Ratio at 120 : Inversed to represent IS}
\item{Gutt_index: Gutt Index : IS index}
\item{Avignon_Si0: Avignon Index at 0 min}
\item{Avignon_Si120: Avignon Index at 120 min}
\item{Avignon_Sim: Avignon Index mean}
\item{Modified_stumvoll: Modified Stumvoll Index}
\item{Stumvoll_Demographics: Stumvoll Index with Demographics}
\item{Glu_Auc_Mean: Mean Glucose AUC}
\item{Insu_Auc_Mean: Mean Insulin AUC}
\item{Matsuda_Auc: Matsuda Index}
\item{BigttSi: BIGTT-Si}
\item{Ifc_inv: Insulin fold change: Inversed to represent IS}
\item{HIRI_inv: Hepatic Insulin Resistance Index: Inversed to represent IS}
}
}
\item{\strong{adipo}:}{
\itemize{
\item{Revised_QUICKI: Revised QUICK Index}
\item{VAI_Men_inv: Visceral Adiposity Index for Men: Inversed to represent IS}
\item{VAI_Women_inv: Visceral Adiposity Index for Women : Inversed to represent IS}
\item{TG_HDL_C_inv: TG to HDL-C ratio converted to IS :Inversed to represent IS}
\item{TyG_inv: TyG Index converted to IS : Inversed to represent IS}
\item{LAP_Men_inv: Lipid Accumulation Product for Men : Inversed to represent IS}
\item{LAP_Women_inv: Lipid Accumulation Product for Women : Inversed to represent IS}
\item{McAuley_index: McAuley Index}
\item{LIRI_inv: Liver Insulin Resistance Index converted to IS : Inversed to represent IS}
\item{Adipo_inv: Adipose Insulin Resistance Index converted to IS : Inversed to represent IS}
\item{Lipo_inv: Lipolysis Index converted to IS : Inversed to represent IS}
\item{ATIRI_inv: Adipose Tissue Insulin Resistance Index converted to IS : Inversed to represent IS}
\item{Belfiore_inv_FFA: Belfiore Index for FFA converted to IS : Inversed to represent IS}
}
}
\item{\strong{tracer_dxa}:}{
\itemize{
\item{LIRI_inv: Liver Insulin Resistance Index converted to IS : Inversed to represent IS}
\item{Lipo_inv: Lipolysis Index converted to IS : Inversed to represent IS}
\item{ATIRI_inv: Adipose Tissue Insulin Resistance Index converted to IS : Inversed to represent IS}
}
}
}

The calculation of most indices follows established formulas documented in the references, with units and other details conforming to the standards set forth in the literature. Although not all original references are explicitly provided, they were consulted individually for each index calculation.

References:
\itemize{
\item Amalia Gastaldelli (2022). (PubMed: \url{https://onlinelibrary.wiley.com/doi/10.1002/oby.23503})
\item Carlos Lorenzo (2010). (PubMed: \url{https://academic.oup.com/jcem/article/95/11/5082/2835314})
}
}
\examples{
data(example_data)
# Example usage of the isi_calculator function
# Run the isi_calculator function with the sample data
# run for each category separately
result <- isi_calculator(example_data, category = "fasting")
result <- isi_calculator(example_data, category = "ogtt")
result <- isi_calculator(example_data, category = "adipo")
result <- isi_calculator(example_data, category = "tracer_dxa")
# OR all four together if you all the required columns
result <- isi_calculator(example_data, category = c("adipo", "ogtt", "fasting", "tracer_dxa"))
# View the results
print(result)
# use ?example_data to see the sample data column names and description

}
\keyword{internal}
